feat(metrics): update metrics when filtering reviews
