+++ client/src/components/Routes/components/Course/Course.container.tsx
@@ -1,15 +1,24 @@
 import React from 'react';
 import { Helmet } from 'react-helmet';
 import { useParams } from 'react-router';
+import { QueryParam } from 'src/core';
+import useQueryParams from 'src/core/hooks/useQueryParams';
+import asArray from 'src/core/utils/asArray';
 import { useCourseQuery } from 'src/graphql';
 
 import Course from './Course';
 
 const CourseContainer: React.FC = () => {
   const { id } = useParams<{ id: string }>();
+  const { semester: semester_ids, difficulty: difficulties } = useQueryParams<{
+    [QueryParam.Semester]: string[];
+    [QueryParam.Difficulty]: string[];
+  }>();
   const { data } = useCourseQuery({
     variables: {
       id,
+      semester_ids: asArray<string>(semester_ids),
+      difficulties: asArray<string>(difficulties).map((_) => Number(_)),
     },
     fetchPolicy: 'no-cache',
   });
